+++ beem/utils.py
@@ -111,7 +111,7 @@ def sanitize_permlink(permlink):
 
 
 def derive_permlink(title, parent_permlink=None, parent_author=None,
-                    max_permlink_length=256):
+                    max_permlink_length=256, with_suffix=True):
     """Derive a permlink from a comment title (for root level
     comments) or the parent permlink and optionally the parent
     author (for replies).
@@ -120,20 +120,38 @@ def derive_permlink(title, parent_permlink=None, parent_author=None,
     suffix = "-" + formatTime(datetime.utcnow()) + "z"
     if parent_permlink and parent_author:
         prefix = "re-" + sanitize_permlink(parent_author) + "-"
-        rem_chars = max_permlink_length - len(suffix) - len(prefix)
+        if with_suffix:
+            rem_chars = max_permlink_length - len(suffix) - len(prefix)
+        else:
+            rem_chars = max_permlink_length - len(prefix)
         body = sanitize_permlink(parent_permlink)[:rem_chars]
-        return prefix + body + suffix
+        if with_suffix:
+            return prefix + body + suffix
+        else:
+            return prefix + body
     elif parent_permlink:
         prefix = "re-"
-        rem_chars = max_permlink_length - len(suffix) - len(prefix)
+        if with_suffix:
+            rem_chars = max_permlink_length - len(suffix) - len(prefix)
+        else:
+            rem_chars = max_permlink_length - len(prefix)
         body = sanitize_permlink(parent_permlink)[:rem_chars]
-        return prefix + body + suffix
+        if with_suffix:
+            return prefix + body + suffix
+        else:
+            return prefix + body
     else:
-        rem_chars = max_permlink_length - len(suffix)
+        if with_suffix:
+            rem_chars = max_permlink_length - len(suffix)
+        else:
+            rem_chars = max_permlink_length
         body = sanitize_permlink(title)[:rem_chars]
         if len(body) == 0:  # empty title or title consisted of only special chars
             return suffix[1:]  # use timestamp only, strip leading "-"
-        return body + suffix
+        if with_suffix:
+            return body + suffix
+        else:
+            return body
 
 
 def resolve_authorperm(identifier):
@@ -347,6 +365,8 @@ def derive_tags(tags):
     elif len(tags.split(" ")) > 1:
         for tag in tags.split(" "):
             tags_list.append(tag.strip())
+    elif len(tags) > 0:
+        tags_list.append(tags.strip())
     return tags_list